add crud all

+++ src/paciente/paciente.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { CreatePacienteDto } from './dto/create-paciente.dto.js';
 import { UpdatePacienteDto } from './dto/update-paciente.dto.js';
 import { PrismaService } from '../prisma/prisma.service.js';
@@ -16,39 +16,57 @@ export class PacienteService {
 
   async findAll() {
     return await this.prisma.paciente.findMany({
-      orderBy: { id: 'asc' },
-      omit: { deleted: true },
       where: { deleted: false },
+      omit: { deleted: true },
+      orderBy: { id: 'asc' },
     });
   }
 
   async findOne(id: number) {
-    return await this.prisma.paciente.findFirst({
+    const paciente = await this.prisma.paciente.findFirst({
       where: { id, deleted: false },
       include: {
         consultas: { omit: { deleted: true } },
       },
       omit: { deleted: true },
     });
+    if (!paciente)
+      throw new NotFoundException(`El ID:${id}, no fue encontrado.`);
+    return paciente;
   }
 
   async update(id: number, updatePacienteDto: UpdatePacienteDto) {
+    const paciente = await this.prisma.paciente.findFirst({
+      where: { id, deleted: false },
+    });
+    if (!paciente)
+      throw new NotFoundException(`El ID:${id}, no fue encontrado.`);
     return await this.prisma.paciente.update({
-      where: { id },
+      where: { id, deleted: false },
       data: updatePacienteDto,
       omit: { deleted: true },
     });
   }
 
   async removeSoft(id: number) {
+    const paciente = await this.prisma.paciente.findFirst({
+      where: { id, deleted: false },
+    });
+    if (!paciente)
+      throw new NotFoundException(`El ID:${id}, no fue encontrado.`);
     return await this.prisma.paciente.update({
-      where: { id },
+      where: { id, deleted: false },
       data: { deleted: true },
       omit: { deleted: true },
     });
   }
 
   async remove(id: number) {
+    const paciente = await this.prisma.paciente.findFirst({
+      where: { id, deleted: false },
+    });
+    if (!paciente)
+      throw new NotFoundException(`El ID:${id}, no fue encontrado.`);
     return await this.prisma.paciente.delete({
       where: { id },
     });
